# src/main/resources/static/swagger/openapi3.yaml
openapi: 3.0.1
info:
  title: ujax API
  description: 백준 알고리즘 스터디 프로젝트 백엔드 API
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
security:
- bearerAuth: []
...
paths:
...
  /api/v1/users/me:
    get:
      tags:
      - User
      summary: 내 정보 조회
      description: 로그인한 사용자의 정보를 조회합니다
      operationId: user-get-me
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse-UserResponse"
              examples:
                user-get-me:
                  value: |-
                    {
                      "success" : true,
                      "data" : {
                        "id" : 1,
                        "email" : "test@example.com",
                        "name" : "테스트유저",
                        "profileImageUrl" : "https://example.com/profile.jpg",
                        "provider" : "GOOGLE",
                        "baekjoonId" : null
                      },
                      "message" : null
                    }
    delete:
      tags:
      - User
      summary: 회원 탈퇴
      description: 로그인한 사용자의 계정을 삭제합니다
      operationId: user-delete-me
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse-Void"
              examples:
                user-delete-me:
                  value: |-
                    {
                      "success" : true,
                      "data" : null,
                      "message" : null
                    }
    patch:
      tags:
      - User
      summary: 내 정보 수정
      description: 로그인한 사용자의 정보를 수정합니다
      operationId: user-update-me
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
            examples:
              user-update-me:
                value: |-
                  {
                    "name" : "수정된이름",
                    "profileImageUrl" : "https://example.com/new-profile.jpg",
                    "baekjoonId" : null
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse-UserResponse"
              examples:
                user-update-me:
                  value: |-
                    {
                      "success" : true,
                      "data" : {
                        "id" : 1,
                        "email" : "test@example.com",
                        "name" : "수정된이름",
                        "profileImageUrl" : "https://example.com/new-profile.jpg",
                        "provider" : "GOOGLE",
                        "baekjoonId" : null
                      },
                      "message" : null
                    }
...
  /api/v1/users/me/profile-image/presigned-url:
    post:
      tags:
      - User
      summary: 프로필 이미지 업로드 Presigned URL 생성
      description: "S3에 프로필 이미지를 업로드하기 위한 Presigned URL을 생성합니다. JPEG, PNG, WEBP만 허\
        용되며 최대 5MB입니다."
      operationId: user-create-profile-image-presigned-url
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/ProfileImageUploadRequest"
            examples:
              user-create-profile-image-presigned-url:
                value: |-
                  {
                    "contentType" : "image/png",
                    "fileSize" : 1048576
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse-PresignedUrlResponse"
              examples:
                user-create-profile-image-presigned-url:
                  value: |-
                    {
                      "success" : true,
                      "data" : {
                        "presignedUrl" : "https://ujax-profile-images.s3.ap-northeast-2.amazonaws.com/presigned?X-Amz-Algorithm=...",
                        "imageUrl" : "https://ujax-profile-images.s3.ap-northeast-2.amazonaws.com/users/1/profile/uuid.png"
                      },
                      "message" : null
                    }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "로그인 후 받은 accessToken을 입력하세요"
  schemas:
...
    ProfileImageUploadRequest:
      title: ProfileImageUploadRequest
      required:
      - contentType
      - fileSize
      type: object
      properties:
        fileSize:
          type: number
          description: "파일 크기 (바이트, 최대 5MB)"
        contentType:
          type: string
          description: "이미지 Content-Type (image/jpeg, image/png, image/webp)"
...
    ApiResponse-UserResponse:
      title: ApiResponse-UserResponse
      required:
      - data
      - success
      type: object
      properties:
        data:
          required:
          - email
          - id
          - name
          - provider
          type: object
          properties:
            provider:
              type: string
              description: "인증 제공자 (GOOGLE, KAKAO, LOCAL)"
            name:
              type: string
              description: 이름
            baekjoonId:
              type: string
              description: 백준 아이디
              nullable: true
            id:
              type: number
              description: 유저 ID
            profileImageUrl:
              type: string
              description: 프로필 이미지 URL
              nullable: true
            email:
              type: string
              description: 이메일
          description: 응답 데이터
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 메시지
          nullable: true
...
    ApiResponse-Void:
      title: ApiResponse-Void
      required:
      - data
      - success
      type: object
      properties:
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 메시지
          nullable: true
...
    ApiResponse-PresignedUrlResponse:
      title: ApiResponse-PresignedUrlResponse
      required:
      - data
      - success
      type: object
      properties:
        data:
          required:
          - imageUrl
          - presignedUrl
          type: object
          properties:
            imageUrl:
              type: string
              description: 업로드 완료 후 이미지 접근 URL
            presignedUrl:
              type: string
              description: S3 업로드용 Presigned URL
          description: 응답 데이터
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 메시지
          nullable: true
...
    UserUpdateRequest:
      title: UserUpdateRequest
      type: object
      properties:
        name:
          type: string
          description: 수정할 이름 (30자 이내)
          nullable: true
        baekjoonId:
          type: string
          description: 백준 아이디
          nullable: true
        profileImageUrl:
          type: string
          description: 수정할 프로필 이미지 URL
          nullable: true